Make use of an external FXML to create the links
Instead of creating the titledPane containing the link and its
settings from scratch, it make it possible to create it based
on an existing FXML file representing the structure of a link.

It make it easy to create the view setting of the links without
needed to build it completely with the java code.