Bind links to their representations in the view.
Gets more informations from youtube about the links given by the user.
Uses the Jackson API to Parse these informations and binds them to
their corresponding links in the linklist.
Binds each link to its view representation in the linkview thus each
interaction mades on the view will be replicated on the backend to the
correspondent link.