Reorganize some classes.
Correct use of the class EmployeeGetMoreInformation in the class
LinksViewContoller.
Review the class AddDialogViewController to make it possible
to show the index of each link in the view without modifying the
url in the linkList.
This information (show the index) is necessary in order to provide
the user during the deletion operation with the position of each link
in the view.