Delete the addToLinksList(String) method.
This method was not working properly and does not
have to be there. The purpose of this was to able
the user added just one link by typed ";;" but the
call to the method was not always consistent with the
code runing in the initializeBinding method.
the use of ";;" was showing to the user an index url x+1
instead of x.