Use of javafx Task and Service class
Using Task and Service classes make the concurrency in the app
easier.
The Service start the worker/task that have to be done
in one or more background thread but still communicate with the
Javafx Application Thread. Thus it is easy to update the UI whenever
it is necessary to give a feedback to the user or response to the
request.