Provide AddLinkService and rename Download
Uses AddLinkService to delegate the addLink method to a Worker.
Using this service makes it possible to keep interacting with
the UI when the task is running in background upon a thread. At
the end of this task, the service provides a list of items to
add to the view. Because the class Download was a service too I
found it logic to append rename it to DownloadService to clarify
that it is not just a class but a service class and thus can run
a task in background.