Improve the UI and replace some control
Instead of using the ProgressBar use the ProgressIndicator to
get more place on the task bar to add a field to notified the
user about what is going on.
A ProgressIndicator was added on the main view port to let the
user know when some information about the links he gave are
retrieving from Youtube.